requirements.txt
easyocr
Pillow
easyocr
streamlit
pandas
numpy